requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.25.0
datasets>=2.12.0
numpy>=1.21.0
pyyaml>=6.0
torchinfo
